refactor: move add_checkpoint to AbstractDataContext
